+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Yii Database Change Log
 
-## 1.3.1 under development
+## 2.0.0 under development
 
 - Enh #815: Refactor `Query::column()` method (@Tigrov) 
+- Enh #766: Allow `ColumnInterface` as column type. (@Tigrov)
 
 ## 1.3.0 March 21, 2024